feat(user can add to fridge with quantity)

+++ server/db/models/Fridge.js
@@ -2,7 +2,7 @@ const Sequelize = require('sequelize');
 const db = require('../db');
 
 const Fridge = db.define('fridge', {
-  quantity: Sequelize.INTEGER,
+  quantity: Sequelize.FLOAT,
 
 });
